requirements.txt
# Core dependencies
dbt-core==1.7.10
dbt-duckdb==1.7.4
pandas==2.2.1
duckdb==0.10.0
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.31.1
pytest==8.0.2

# Optional warehouse adapters (uncomment as needed)
# dbt-postgres==1.7.10
# dbt-snowflake==1.7.3
# dbt-bigquery==1.7.7

# Additional dependencies
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0